data_extraction/requirements.txt
google-generativeai>=0.3.0
mongoengine>=0.27.0
python-dotenv>=1.0.0
retrying>=1.3.4
pytest>=7.3.1
python-docx
colorlog
scikit-learn
pandas
matplotlib
pandoc
seaborn
colorama
graphviz
langchain